requirements.txt
transformers>=4.41,<5.0
torch>=2.0,<3.0
pandas>=2.0
numpy>=1.26
python-dotenv>=1.0
pytest>=7.4
huggingface_hub>=0.20
